Pass onClear function as prop to FilterSummary component
