+++ src/domain/collection/eventList/EventList.tsx
@@ -9,8 +9,8 @@ import useLocale from '../../../hooks/useLocale';
 import getLocalisedString from '../../../util/getLocalisedString';
 import Container from '../../app/layout/Container';
 import { EVENT_SORT_OPTIONS, PAGE_SIZE } from '../../eventSearch/constants';
-import { getEventFilters, getNextPage } from '../../eventSearch/EventListUtils';
 import EventSearchList from '../../eventSearch/searchResultList/EventList';
+import { getEventFilters, getNextPage } from '../../eventSearch/utils';
 import styles from './eventList.module.scss';
 
 interface Props {
