+++ src/domain/collection/eventList/__tests__/EventList.test.tsx
@@ -11,7 +11,7 @@ import { SUPPORT_LANGUAGES } from '../../../../constants';
 import { EventListDocument } from '../../../../generated/graphql';
 import mockEvent from '../../../event/__mocks__/eventDetails';
 import { EVENT_SORT_OPTIONS, PAGE_SIZE } from '../../../eventSearch/constants';
-import { getEventFilters } from '../../../eventSearch/EventListUtils';
+import { getEventFilters } from '../../../eventSearch/utils';
 import EventList from '../EventList';
 
 const searchParams = new URLSearchParams(
