+++ src/domain/eventSearch/EventSearchPageContainer.tsx
@@ -10,10 +10,10 @@ import useIsSmallScreen from '../../hooks/useIsSmallScreen';
 import useLocale from '../../hooks/useLocale';
 import PageWrapper from '../app/layout/PageWrapper';
 import { EVENT_SORT_OPTIONS, PAGE_SIZE } from './constants';
-import { getEventFilters, getNextPage } from './EventListUtils';
 import styles from './eventSearchPage.module.scss';
 import Search from './Search';
 import SearchResultList from './searchResultList/SearchResultList';
+import { getEventFilters, getNextPage } from './utils';
 
 const EventSearchPageContainer: React.FC<RouteComponentProps> = () => {
   const { t } = useTranslation();
