+++ src/domain/landingPage/landingPageSearch/LandingPageSearch.tsx
@@ -21,8 +21,9 @@ import IconMusic from '../../../icons/IconMusic';
 import IconSports from '../../../icons/IconSports';
 import IconTheatre from '../../../icons/IconTheatre';
 import IconTree from '../../../icons/IconTree';
-import { getSearchQuery } from '../../../util/searchUtils';
 import { ROUTES } from '../../app/constants';
+import { DEFAULT_SEARCH_FILTERS } from '../../eventSearch/constants';
+import { getSearchQuery } from '../../eventSearch/utils';
 import styles from './landingPageSearch.module.scss';
 
 const Search: FunctionComponent = () => {
@@ -37,17 +38,11 @@ const Search: FunctionComponent = () => {
 
   const handleCategoryClick = (category: Category) => {
     const search = getSearchQuery({
+      ...DEFAULT_SEARCH_FILTERS,
       categories: [category.value],
       dateTypes,
-      divisions: [],
       end,
-      isFree: false,
-      keywordNot: [],
-      keywords: [],
-      places: [],
-      publisher: null,
       start,
-      text: '',
     });
 
     push({ pathname: `/${locale}${ROUTES.EVENTS}`, search });
